+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 3.0.0 (July 15, 2022)
 * Updated method `uploadAttachment` from `AttachmentProcessor`
+* Added method `getMaesterAttachmentUrlById` which return url to Maester attachment
 
 ## 2.1.0 (May 06, 2022)
 * Now method `uploadAttachment` from `AttachmentProcessor` automatically sets the `contentType`